move example into module
