+++ Dockerfile
@@ -1,4 +1,5 @@
 # adapted from https://dev.to/farcellier/package-a-poetry-project-in-a-docker-container-for-production-3b4m
+# and  https://medium.com/@albertazzir/blazing-fast-python-docker-builds-with-poetry-a78a66f5aed0
 
 FROM python:3.12.5-slim-bookworm AS build
 
@@ -22,4 +23,4 @@ COPY athena_federation ./athena_federation
 # Only needed if you, e.g., install a script
 RUN poetry install --without dev 
 
-ENTRYPOINT ["poetry", "run", "python", "-m", "athena_federation.main"]
+ENTRYPOINT ["poetry", "run", "python", "-m", "tests.handler.lambda_handler"]
